remove used casts

+++ src/main/java/net/optifine/LightMap.java
@@ -102,8 +102,8 @@ public class LightMap
 
     private void getLightMapColumn(float[][] origMap, float x, int offset, int width, float[][] colRgb)
     {
-        int i = (int)Math.floor((double)x);
-        int j = (int)Math.ceil((double)x);
+        int i = (int)Math.floor(x);
+        int j = (int)Math.ceil(x);
 
         if (i == j)
         {
